refactor(analytics): relocate getTimeRangeFilter utility to elasticsearch/utils and remove the obsolete file
